moved logic of errors to another file
